@@ -9,14 +9,14 @@ const double kDefaultIconSize = 16;
 // const double kDefaultMenuFont =
 const double kDefaultMenuFontSize = 12;
 const String kProgramName = 'fnmap';
-const String kAppVersion = '1.1-alpha-1';
+const String kAppVersion = '1.3-1-1';
 const String kPackageName = 'com.krioltech.fnmap';
 const String kProfileFilename = 'scan_profile.usp';
 const String kConfigFilename = 'fnmap.conf';
 const String kZenmapConfFilename = 'zenmap.conf';
 const String kCustomKey = 'Custom';
 //const Color kDefaultColor = Colors.teal;
-const Color kDefaultColor = Colors.indigo;
+const Color kDefaultColor = Colors.white70;
 const Color kAccentColor = Colors.indigoAccent;
 const Color kValidColor = Color(0xff2e7d32);  // Green
 const Color kInvalidColor = Color(0xffdd2c00); // Red;
